move tasks and notes
