+++ RepairsApi/V2/UseCase/ListVariationTasksUseCase.cs
@@ -5,6 +5,7 @@ using RepairsApi.V2.UseCase.Interfaces;
 using System;
 using RepairsApi.V2.Gateways;
 using System.Linq;
+using RepairsApi.V2.Factories;
 
 namespace RepairsApi.V2.UseCase
 {
@@ -18,19 +19,7 @@ namespace RepairsApi.V2.UseCase
         public async Task<IEnumerable<VariationTasksModel>> Execute(int workOrderId)
         {
             var tasks = await _jobStatusUpdateGateway.SelectWorkOrderVariationTasks(workOrderId);
-
-            return tasks.Select(t => new VariationTasksModel
-            {
-
-                Id = t.Id.ToString(),
-                Code = t.CustomCode,
-                Description = t.CustomName,
-                UnitCost = t.CodeCost,
-                OldQuantity = t.OriginalQuantity,
-                NewQuantity = t.Quantity?.Amount
-            }
-            ).ToList();
-
+            return tasks.Select(t => t.ToVariationsResponse()).ToList();
         }
     }
 }
